+++ src/Ace/Drawable.cpp
@@ -168,6 +168,8 @@ namespace ace
 
                         // Update
                         // TODO: Flip
+
+						sprite.SetID(tile.ID);
                         sprite.Texcoord(sheet.GetSprite(tile.ID)->texcoord);
                         sprite.Scale(Vector2(tilesetSize.x, tilesetSize.y) / math::Min(tileSize.x, tileSize.y));
 
@@ -190,7 +192,7 @@ namespace ace
                     }
                 }
 
-
+				GraphicsDevice::CreateIndicies(layer.indexBuffer, layer.tiles.size(), BufferUsage::Static);
                 layers.push_back(layer);
             }
         }
@@ -237,11 +239,13 @@ namespace ace
 
     void Tilemap::TileLayer::Draw() const
     {
-        for (UInt32 i = 0; i < tiles.size(); ++i)
-        {
-            // TODO: Draw all sprites. (single draw call)
-            GraphicsDevice::Draw(tiles[i]);
-        }
+		GraphicsDevice::Draw(tiles.data(), tiles.size(), indexBuffer);
+
+        //for (UInt32 i = 0; i < tiles.size(); ++i)
+        //{
+        //    // TODO: Draw all sprites. (single draw call)
+        //    GraphicsDevice::Draw(tiles[i]);
+        //}
     }
 
     tmx::Map& Tilemap::GetMap()